v.0.4.0

+++ src/discordEvents/chat.ts
@@ -1,7 +1,8 @@
 import { Event } from "."
 import { format } from "../format"
-import { config, globals, logger, saveAllUsers } from "../index"
-import { sendToServer } from "../server"
+import { globals, logger, saveAllUsers } from "../index"
+import { sendCommand, sendToServer } from "../server"
+import { config } from "../config"
 
 const delay = (ms: number) => new Promise((resolve) => setTimeout(resolve, ms))
 
@@ -14,9 +15,23 @@ const online: Event = {
 	callback: async (message) => {
 		if (globals.serverStarted) {
 			console.log(message)
+			var msg = message.content
+			if (message.mentions.users.size > 0) {
+				for (const [key, user] of message.mentions.users.entries()) {
+					msg = msg.replace(
+						new RegExp(`<@.${user.id}>`, "g"),
+						"@" + (await message.guild?.members.fetch(user))?.nickname ?? user.username
+					)
+				}
+			}
+			if (message.mentions.roles.size > 0) {
+				message.mentions.roles.forEach((role) => {
+					msg = msg.replace(new RegExp(`<@.${role.id}>`, "g"), "@" + role.name)
+				})
+			}
 			sendToServer(
 				format(config.FactorioMessageFormat, {
-					message: message.content.trim(),
+					message: msg.trim(),
 					user: message.member?.nickname ?? message.author.username,
 				})
 			).catch((err) => {
@@ -28,6 +43,7 @@ const online: Event = {
 						})
 					)
 					.then(async (message) => {
+						if (config.Errors.DeleteTimeout <= 0) return
 						await delay(config.Errors.DeleteTimeout * 1000)
 						message.delete()
 					})
@@ -43,12 +59,13 @@ const online: Event = {
 					})
 				)
 				.then(async (message) => {
+					if (config.Errors.DeleteTimeout <= 0) return
 					await delay(config.Errors.DeleteTimeout * 1000)
 					message.delete()
 				})
 			return
 		}
-	},
+	}
 }
 
 export default online
